+++ test/dart_api/value/null_test.dart
@@ -24,6 +24,7 @@ void main() {
 
   test("isn't any type", () {
     expect(value.assertBoolean, throwsSassScriptException);
+    expect(value.assertCalculation, throwsSassScriptException);
     expect(value.assertColor, throwsSassScriptException);
     expect(value.assertFunction, throwsSassScriptException);
     expect(value.assertMap, throwsSassScriptException);