+++ test/dart_api/value/number_test.dart
@@ -130,6 +130,7 @@ void main() {
 
     test("isn't any other type", () {
       expect(value.assertBoolean, throwsSassScriptException);
+      expect(value.assertCalculation, throwsSassScriptException);
       expect(value.assertColor, throwsSassScriptException);
       expect(value.assertFunction, throwsSassScriptException);
       expect(value.assertMap, throwsSassScriptException);