+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 7.2.0
+
+* No user-visible changes.
+
 ## 7.1.6
 
 * No user-visible changes.